solve simple compiler warnings

+++ src/p_saveg.c
@@ -34,7 +34,7 @@ byte *save_p;
 
 // Pads save_p to a 4-byte boundary
 //  so that the load/save works on SGI&Gecko.
-#define PADSAVEP() save_p += (4 - ((int)save_p & 3)) & 3
+#define PADSAVEP() save_p += (4 - ((long int)save_p & 3)) & 3
 
 //
 // P_ArchivePlayers
@@ -85,7 +85,7 @@ void P_UnArchivePlayers(void) {
     for (j = 0; j < NUMPSPRITES; j++) {
       if (players[i].psprites[j].state) {
         players[i].psprites[j].state =
-            &states[(int)players[i].psprites[j].state];
+            &states[(long int)players[i].psprites[j].state];
       }
     }
   }
@@ -255,10 +255,10 @@ void P_UnArchiveThinkers(void) {
       mobj = Z_Malloc(sizeof(*mobj), PU_LEVEL, NULL);
       memcpy(mobj, save_p, sizeof(*mobj));
       save_p += sizeof(*mobj);
-      mobj->state = &states[(int)mobj->state];
+      mobj->state = &states[(long int)mobj->state];
       mobj->target = NULL;
       if (mobj->player) {
-        mobj->player = &players[(int)mobj->player - 1];
+        mobj->player = &players[(long int)mobj->player - 1];
         mobj->player->mo = mobj;
       }
       P_SetThingPosition(mobj);
@@ -430,7 +430,7 @@ void P_UnArchiveSpecials(void) {
       ceiling = Z_Malloc(sizeof(*ceiling), PU_LEVEL, NULL);
       memcpy(ceiling, save_p, sizeof(*ceiling));
       save_p += sizeof(*ceiling);
-      ceiling->sector = &sectors[(int)ceiling->sector];
+      ceiling->sector = &sectors[(long int)ceiling->sector];
       ceiling->sector->specialdata = ceiling;
 
       if (ceiling->thinker.function.acp1)
@@ -445,7 +445,7 @@ void P_UnArchiveSpecials(void) {
       door = Z_Malloc(sizeof(*door), PU_LEVEL, NULL);
       memcpy(door, save_p, sizeof(*door));
       save_p += sizeof(*door);
-      door->sector = &sectors[(int)door->sector];
+      door->sector = &sectors[(long int)door->sector];
       door->sector->specialdata = door;
       door->thinker.function.acp1 = (actionf_p1)T_VerticalDoor;
       P_AddThinker(&door->thinker);
@@ -456,7 +456,7 @@ void P_UnArchiveSpecials(void) {
       floor = Z_Malloc(sizeof(*floor), PU_LEVEL, NULL);
       memcpy(floor, save_p, sizeof(*floor));
       save_p += sizeof(*floor);
-      floor->sector = &sectors[(int)floor->sector];
+      floor->sector = &sectors[(long int)floor->sector];
       floor->sector->specialdata = floor;
       floor->thinker.function.acp1 = (actionf_p1)T_MoveFloor;
       P_AddThinker(&floor->thinker);
@@ -467,7 +467,7 @@ void P_UnArchiveSpecials(void) {
       plat = Z_Malloc(sizeof(*plat), PU_LEVEL, NULL);
       memcpy(plat, save_p, sizeof(*plat));
       save_p += sizeof(*plat);
-      plat->sector = &sectors[(int)plat->sector];
+      plat->sector = &sectors[(long int)plat->sector];
       plat->sector->specialdata = plat;
 
       if (plat->thinker.function.acp1)
@@ -482,7 +482,7 @@ void P_UnArchiveSpecials(void) {
       flash = Z_Malloc(sizeof(*flash), PU_LEVEL, NULL);
       memcpy(flash, save_p, sizeof(*flash));
       save_p += sizeof(*flash);
-      flash->sector = &sectors[(int)flash->sector];
+      flash->sector = &sectors[(long int)flash->sector];
       flash->thinker.function.acp1 = (actionf_p1)T_LightFlash;
       P_AddThinker(&flash->thinker);
       break;
@@ -492,7 +492,7 @@ void P_UnArchiveSpecials(void) {
       strobe = Z_Malloc(sizeof(*strobe), PU_LEVEL, NULL);
       memcpy(strobe, save_p, sizeof(*strobe));
       save_p += sizeof(*strobe);
-      strobe->sector = &sectors[(int)strobe->sector];
+      strobe->sector = &sectors[(long int)strobe->sector];
       strobe->thinker.function.acp1 = (actionf_p1)T_StrobeFlash;
       P_AddThinker(&strobe->thinker);
       break;
@@ -502,7 +502,7 @@ void P_UnArchiveSpecials(void) {
       glow = Z_Malloc(sizeof(*glow), PU_LEVEL, NULL);
       memcpy(glow, save_p, sizeof(*glow));
       save_p += sizeof(*glow);
-      glow->sector = &sectors[(int)glow->sector];
+      glow->sector = &sectors[(long int)glow->sector];
       glow->thinker.function.acp1 = (actionf_p1)T_Glow;
       P_AddThinker(&glow->thinker);
       break;
